Add Makefile to deploy bosh-lite on Virtualbox
Add a Makefile that follows the instructions described in [1] and [2]

We generate a local manifest and store it, with the credentials, in
the directory generated. We can commit that, because this is an
environment in virtualbox, or just delete it.

There is also a target to configure the environment alias for bosh
and the route.

[1] https://bosh.io/docs/bosh-lite
[2] http://www.starkandwayne.com/blog/bosh-lite-on-virtualbox-with-bosh2/