Add target to deploy bosh on AWS
We will deploy bosh using create-env with bosh2.

We use the default example IPs, hardcoded for the time being.
We configure an instance profile, although empty for the time being.

The agent, director and SSH would be accessible, secured with
security groups. In the future we would use tunnels or so.

In order to be able to use STS credentials to deploy, we
must use a custom release of bosh-aws-cpi, as per the PR
https://github.com/cloudfoundry-incubator/bosh-aws-cpi-release/pull/70