Expose nginx reverse proxy ports 80+443
We want to access the cf-router from internet. We will expose
the nginx reverse proxy with basic auth using the port_forwarding
job from the networking release[1].

We must give the nginx a static IP, e.g. 10.244.3.34 so we can
configure the port_forwarding.

We add a new security group to allow access to these ports from
the admin IPs.