Configure a public DNS for the public access
We configure a public wildcard dns pointing to the bosh-lite public
IP, so that we can access via HTTP/HTTPS.

We allow configure the root zone in the Makefile, passing the
variable to terraform.

We want to access the prometheus domain, so we extend the
route registration to add the public domains.