Another attempt to implement some basicauth
We check for the presence of a auth cookie, if it is not there, we
redirect to a endpoint that would do the basic auth and set that
cookie if the auth is successful.

It kinda works, but I have problems with graphana when loading
plugins:

         Plugin Error
         Type error Instantiating
         https://grafana.bosh-lite-1.sandbox.keytwine.com/public/plugins/grafana-piechart-panel/module.js
         Loading plugins/grafana-piechart-panel/module

It is like not sending the cookie when loading a module.js.

I do not know why this happens, so I will have to disable all of
this and use basic IP restrictions :(