Check if the host header matches the public name
We failed using basicauth for retrict access to the bosh-lite
when open to the public... so we will basically restrict access
based on the Host header, checking that the request matches
the public domain.

This would prevent access to any service that are not the ones
published explicitly under the public domain.